#5 Delete old dogu registry.
